Working TeamInMatchDetails, Graphing, and TeamDetails (#29)
* App Runs (mostly), working on SpecificConstants

* Push for Colin

* Working TeamInMatchDetails, Graphing, and TeamDetails

+++ androidviewertemplates/src/main/java/com/example/evan/androidviewertemplates/firebase_classes/CalculatedTeamData.java
@@ -102,7 +102,7 @@ public class CalculatedTeamData extends Object {
     public Float dysfunctionalPercentage;
 
 
-    public Float number;
+    public Integer number;
     public Float climbPercentage;
     public Integer predictedNumAllianceSwitchCubesAuto;
     public Integer predictedNumScaleCubesAuto;
